migration/cgs: add cgs_mig_is_ready
cgs_mig_is_ready is used to check if cgs migration is ready when required.
This is done by the framwork API calling to each vendor specific function
to check, and the check is performed at the time when user requests
migration to start. For the legacy VM migration case and the vendor
specific implementation that don't require the check, just return true
to have the migration flow continue.

cgs.c file is created to hold all the cgs migration framwork APIs,
and cgs.h exposes the APIs to the migration flow.

Signed-off-by: Wei Wang <wei.w.wang@intel.com>