kvm/tdx: Enable SendMessage for vTPM client
A simply protocol "TRANS_PROTOCOL" is introduced
for data transfering between vTPM client and server,
the main reason to introduce this to pass vTPM ID
from client side to server side.

To simply the implementation the vTPM ID is carried
every time even it's only necessary for 1st time
communicats with the server side, the vTPM
client/server communication frequency is not high.

Signed-off-by: Yao Yuan <yuan.yao@intel.com>